changed size_t types to int to squash warnings

+++ day 8/include/coordinate.hh
@@ -1,20 +1,18 @@
 #ifndef COORDINATE_H
 #define COORDINATE_H
 
-#include <cstddef>
-
 class Coordinate {
 
   public:
 
-    Coordinate(size_t x, size_t y);
+    Coordinate(int x, int y);
     Coordinate(const Coordinate& other);
     Coordinate& operator=(const Coordinate& other);
     Coordinate operator-(const Coordinate& other) const;
     bool operator==(const Coordinate& other) const;
 
-    size_t x_ = 0;
-    size_t y_ = 0;
+    int x_ = 0;
+    int y_ = 0;
 
 };
