+++ day 8/include/grid.hh
@@ -1,7 +1,6 @@
 #ifndef _GRID_HH_
 #define _GRID_HH_
 
-#include <cstddef>
 #include <vector>
 
 #include "antenna.hh"
@@ -11,17 +10,17 @@ class Grid {
   public:
 
     Grid();
-    Grid(const std::vector<Antenna>& antennae, const size_t max_x, const size_t max_y);
+    Grid(const std::vector<Antenna>& antennae, const int max_x, const int max_y);
     ~Grid() = default;
-    const size_t GetMaxX() const { return max_x_; }
-    const size_t GetMaxY() const { return max_y_; }
+    const int GetMaxX() const { return max_x_; }
+    const int GetMaxY() const { return max_y_; }
     const std::vector<Antenna>& GetAntennae() const {return antennae_; }
 
   private:
 
     std::vector<Antenna> antennae_;
-    size_t max_x_ = 0;
-    size_t max_y_ = 0;
+    int max_x_ = 0;
+    int max_y_ = 0;
 
 };
